"5:45pm. Let me back this up here.
```
Even if I check the scheme (which is `git:/` in this case,) the problem is that it is fetching the stale version of the file from version control. If you look at the documentation for this function it states that...

```ts
/**
 * An event that is emitted when a [text document](#TextDocument) is opened or when the language id
 * of a text document [has been changed](#languages.setTextDocumentLanguage).
 *
 * To add an event listener when a visible text document is opened, use the [TextEditor](#TextEditor) events in the
 * [window](#window) namespace. Note that:
 *
 * - The event is emitted before the [document](#TextDocument) is updated in the
 * [active text editor](#window.activeTextEditor)
 * - When a [text document](#TextDocument) is already open (e.g.: open in another [visible text editor](#window.visibleTextEditors)) this event is not emitted
 *
 */
export const onDidOpenTextDocument: Event<TextDocument>;
```

More specifically it says that `The event is emitted before the document is updated in the active text editor`. Am I supposed to parse the uri to extract the actual file system path of the document and load the file separately? That seems very unlikely. There should be some function or event to get the text of the actual document being opened directly.

It is not good to me if the text of the document in that event is from a commit a few hours ago.
```

Actually, I do not feel like asking right now. Let me see if there are any events related to the active text editor.

```
export function asRelativePath(pathOrUri: string | Uri, includeWorkspaceFolder?: boolean): string;
```

Maybe I should be using this. Let me see what passing an URI to this does.

```
rel path: c:\Users\Marko\Source\Repos\The Spiral Language\VS Code Plugin\spiral\src\a.spi.git
```

Nothing good. This is extremely strange.

5:55pm. Even though the bot closed the issue, I've made another post in it. Maybe the dev will answer it.

If not, my plan B will be to ask around in the slack channel.

6:20pm. I figured it out thanks to the dev replying. That git file is something the git extension triggers and I should ignore it. The reason why onOpen is not triggering when I launch the extension is because the extension is not active at that point.

With this knowledge in tow, I'll be able to make changes work tomorrow. But tomorrow I'll take a short break to do the monthly review. Now let me shot demons."