"1:25pm. Let me rest for a bit longer.
1:30pm. Just 5-10m more and I will start.

https://github.com/microsoft/vscode/issues/

The stuff in issue 92789 is interesting. It seems that range provider can have issues with flickering since the editor needs to wait for the server to provide the changes for the viewport.

But I guess I'll start off with it regardless.

Now that I am replaning the server, I see that keeping track of the changes would get more complicated.

2:10pm. At first I wanted to slack, but now I am in thought again.

4:55pm. Holy shit, was I just thinking in bed this whole time? It seems so.

All this reminds of me of 2017. I see the language stuff as easy now, but the currency stuff is just as hard to me now as the language stuff was back then.

5pm. Concurrency just requires different patterns of thought, and they are not ones I am familiar with. It took me a really long time to become good at Spiral, both at making it and programming in it. Concurrency is one of the great challenges in programming - it won't be an exception in terms of how much effort it will take to master.

5:05pm. I am not sure I feel like doing anything more in the day.

Since I am not going to be doing programming, let me write down some of the ideas that I had.

I guess the main idea that I had in order to simplify things is that I should ditch the router socket.

One thing I've realized is - I am not really working on an asynchronous server. The reason I am using ZeroMQ in the first place is so I can connect the Node.js parts to the ones on the .NET side. If VS Code was running on the later, I would never even bother with it - it would be all Hopac and synchronous communication across the board.

While I was thinking, I also became completely perplexed. I have absolutely no idea how for example the router or the dealer sockets decide to drop messages. I am sort of thinking that the router is unreliable, but once I try to probe into that, I realize that I do not really understand why I think that.

It is true that I run into issues with them, but are router and dealer really more reliable than request and response?

As I probe deeper it feels, well it feels like I am going senile because I have no idea why I believe the later are synchronous and the former are not.

> When you use REQ to talk to REP, you get a strictly synchronous request-reply dialog. The client sends a request. The service reads the request and sends a reply. The client then reads the reply.

Is this really true? I mean, if I send a request to a router socket...well, I am guessing the send will be a success if it ends up in the other side's buffer.

But I think the guide mentioned that request sockets have a send buffer...

> ZeroMQ uses the concept of HWM (high-water mark) to define the capacity of its internal pipes. Each connection out of a socket or into a socket has its own pipe, and HWM for sending, and/or receiving, depending on the socket type. Some sockets (PUB, PUSH) only have send buffers. Some (SUB, PULL, REQ, REP) only have receive buffers. Some (DEALER, ROUTER, PAIR) have both send and receive buffers.

Ah, Req and Rep only have receive buffers. That means their sends must necessarily be synchronous. And for dealer sockets, they will be asynchronous as it will be putting the messages in the send buffer if it cannot send them right away. Of course, receives will be synchronous.

5:20pm. Right, so my hazy memory of dealer/router being async and the req/res being sync is not wrong.

And in this case, I should be using the later.

I am not doing a stateleess service here - now that the issue is doing the least amount of work needed by sending deltas, missed messages will cause the editor and the language server to desynchronize. Once that happens, the user will get mad and the only choice will be to restart the server.

I want to avoid that. And if something bad happens without the chance to recover rather the server crash loudly.

So...

```fs
        let msg = sock.ReceiveMultipartMessage(3)
        let address = msg.Pop()
        msg.Pop() |> ignore
```

Rather than this thing where I receive a multipart message...

```fs
let uri_editor = "tcp://localhost:13806"
let uri_lang_serv = "tcp://*:13805"
```

I am going to have it connect to a different port and send stuff from the later stages there.

Reliability aside, I am trying to figure out how to make use of a dealer socket to do my stuff on the editor side and am coming up empty. It should be possible, but I really am struggling a lot to simply things in my head. Everything feels awkward and complicated, and it is hard to make things flow naturally.

5:35pm. Yeah, I do not need all this complication. Forget about the server, and just think of it as gluing two processes together.

In that case, doing two binds like the above and making things synchronous as much as possible is a good approach to take. In my thoughts I often indulge myself thinking how I can extend my limits, but here I should be taking the exactly opposite approach. Somewhat ironically, the pattern that I will use here was not in the guide, which why I overlooked it until now.

I learned all this advanced stuff that I missed the value of the simpler patterns.

If I can limit the way I do interprocess communication, that will make other things much easier.

That is what I need to focus on.

5:40pm. Yeah, I do not think I can use a dealer socket on the editor side. Async sends are not good here - I do in fact need confirmation that the request has been received otherwise I might get wrong deltas back.

...Oh, I should note that pair sockets are also async since they have send buffers. In the past I've been tempted to skip the acknowledgement part - in some cases I just need the send to be synchronous, but even if they worked for more than inprocess communication, they are no good for what I want as their sends are async.

5:45pm. I admit, I've been affected - I read that router and dealer are more useful, but haven't exactly thought about why that is. If my Hopac experience is any indication, synchronous messaging is in fact more useful and fundamental.

It is easy to turn synchronous to asynchronous messaging just by putting a buffer between the send and a receive, but going the other way is much harder.

5:50pm. Let me stop here for the day. I really did not think things would go like this today. That I would not realize that only sending individual line changes is a poor idea shows how bad I am still at this. I am still overwhelmed by complexity.

It is going to take some hard though to overcome this. Tomorrow, I will start work on redesigning the servers. Literally none of the pices that I have in mind are hard, but I am struggling to fit them together."