"3:45pm. Just putting a few comments in places.
3:50pm. Now, I need to take some time to think about what is next.

This stupid issue took my whole day and I haven't even dealt with it properly.

3:55pm. Ok, let me step away from the screen for a while to think about the next step. Up to now my attention was wholly on tokenization, and I hadn't primed my brain for anything else today.

At first glance the next step is parsing, but not quite - parsing is the step after the next one.

The next step is blockization.

It is a really simple idea, but the block server will be the supervisor of the whole pipeline. It is what is going to allow me to implement the later stages just as conveytor belts that take in blocks. It is the piece that will chain different packages together into a single thread that the later stages can digest.

It is also going to be the server that will do all the communication with the editor.

I've been imagining using modulation through intermediate nodes and such, but that is a bad idea. After much thinking I finally realized that after typechecking (and the rest of the stages) are done, they need to check back with the supervisor to make sure that the block has not changed. I can't just send whatever comes out of the other end without doing that first.

4:05pm. Let just step away from the screen. When I want to think I should just do it in bed. When I am inactive in front of the screen there is the temptation to surf my usual haunts.

4:55pm. I took that nap for a bit.

Let me make the block parse function. It is really extremely simple. After that I'll call it a day. Today it seems I just did not feel like doing much.

5:25pm.

```fs
module Spiral.Blockize
open Tokenize

/// Groups the comments with the statement following them.
let blocks (ar : LineToken ResizeArray) : Result<LineToken [] [], int * LineToken> =
    let indent_init = 0
    let blocks = ResizeArray()
    let blocks_add (b : _ ResizeArray) = blocks.Add(b.ToArray())
    let rec loop_initial (b : _ ResizeArray) i =
        if i < ar.Count then
            let r,t as x = ar.[i]
            if r.from = indent_init then
                match t with
                | TokComment _ -> b.Add x; loop_initial b (i+1)
                | _ -> b.Add x; loop_body b (i+1)
            else Error(i,x)
        else blocks_add b; Ok(blocks.ToArray())
    and loop_body (b : _ ResizeArray) i =
        if i < ar.Count then
            let r,_ as x = ar.[i]
            if r.from = indent_init then blocks_add b; loop_initial (ResizeArray()) i
            else b.Add x; loop_body b (i+1)
        else blocks_add b; Ok(blocks.ToArray())
    loop_initial (ResizeArray()) 0
```

This is not a bad way of starting this thing off.

5:30pm. The server version of this will essentially do the same thing as the above, except incrementally.

5:35pm. That is what I am going to focus on tomorrow.

Today, let me call it a day early for once. I'll use this time to charge up and fill myself with motivation for what will come next. When the next thing comes I always do not feel like it. That is the wrong way to think. What I need to think is that everything in my life is wrong because I do not have blockization. I need to feel the dissatisfaction over that. Then the piece will get done."