"8:45am. Got a reply to that issue, like half an hour ago. Instead of slacking let me get to it.
9:10am. Let me do my morning reading and then I will start.

9:55am. I am doing some back and forth with the guy in that issue. It turns out the sequential edits I am doing are illegal. This is nowhere documented and it is a bad show from their side. Right now, I do not know how to optimize those sequential edits into a single non-overlapping one.

10:05am. I guess I'll have to do some thinking. Rather than working on blockization, I'll try and see whether I can make a program that optimizes sequential edits into non-overlapping ones.

Let me do the chores here.

10:30am. Done with chores. I am still thinking about that problem.

If it were just insertions, simulating the end result of that for some arbitrary array is not too hard.

Given a set of insertion operations, it is easy to merge them together.

Even with deletions sprinkled in, that would also not be hard to calculate.

Mhhhhh...what is the effect of deletion operations on other deletion operations?

11:35am. Ok, I did some more posting in issue 101744.

Yeah, I am done with this. I am going to stick with the range provider. The changes one is just too difficult. Transforming overlapping into non-overlapping edits is too hard.

11:40am. As expected, unexpected time wasters will appear in my path towards getting this to work. I wanted to start work on blockization this morning, but I will have to save that for after the break. Let me have it here."