"8:50am. Yesterday I forgot to check the VS Code slack channel for any replies. I'll do that in a bit. Right now, let me just chill.
9:35am. Enough slacking. I had a bit of fun, and now let me dive into my purpose. First off let me check the channel.

...Nothing there. Nevermind that. I guess I won't be compiling VS Code. I won't bother checking that channel anymore.

Honestly, the level of involvement from the devs when I ask them questions has been disappointing throughout. But give those 5k unsolved issues, I can understand their need to manage their time.

9:40am. Forget VS Code. Those sort of things will deal with themselves in due time. It is not my job to fix their shit.

What I should do is focus on the task in front of me - the parser.

9:55am. This will really be a boatload of work. It will take me weeks won't it? And even then I won't be done, I will be working on this for a long time.

Forget finishing this in the next week or two. Forget finishing it by the end of the year. This will take as long as it should.

My weakness is always that I want to get things out of the way to move to bigger and better stuff.

This time just sit down and do it properly.

Spiral v0.2 will be my greatest work, and it will also be an admission of my limits as a human.

Even though the real segment that lies underneath is what trully captures the essence of computation more than any type theory, on the surface it will be an exceedingly plain language.

There are such stories of nerds who become normies, and v0.2 is such an event for Spiral. But well, it needs it. The real segment is just too unusable at scale. I want to program in the simplest way possible in my next attempt at grabbing the Singularity. The really big idea of type systems is top-down reasoning about programs. I just cannot oppose it any longer.

This is something that is as pervasive as the air we breathe in today's programming ecosystem, but in a world where Spiral styled partially evaled type inference became the norm, top-down reasoning would have been absolutely radical.

10am. It is a pity, just imagine how much better the language could turn out had I the power of those future brain cores. The will I have now and the desire, they both pale in comparison to the kind of power the future agents will be able to throw around.

I have to work so much just to get a bare sliver of their power.

10:05am. It will be worth it.

Although becoming a normie is a death sentence for a human, I genuinely think that top-down reasoning is worthy of admiration. So a language being *normal* in this context is hardly something to be ashamed of. I really wish I could say the same about being a normal person.

10:15am. I have to find the power to do this one last time. Doing v0.2 will take me long, even after I finish it, I honestly don't thing I'll find the power to do v0.3. That thing would require an entire runtime. Given how long just this is taking me, just where am I going to find the energy to do that.

Nowhere.

I am going to need the power of self improvement for that. There isn't a single human who can give himself to the world forever without getting something back in return.

10:20am. I exceeded many other languages with v0.09. Bottom up reasoning is amazingly powerful. But top-down reasoning is a power too. And I am going to have to tap into it to get me through the next attempt.

I need this power. Without top-down reasoning, I won't be able to win.

v0.09 and v0.1 are a single design point in the language space, and v0.2 will be a different one. I need to stake my claim on it.

10:25am. Probably the greatest achievement in my life is figuring out the principles of self improvement and ownership. If I were to die tomorrow, somebody courageous will read Simulacrum, think that the kind of self improvement I am proposing is a noble aspiration, read the essay on power, and try to attain it in a realistic manner. No matter how crazy they seem, I haven't been able to find any weaknesses in the concepts.

I am solidly sure they are dominant.

The things I did in v0.09 while they haven't taken the world by storm, they do have a transcendental quality as well. A lot of the techniques I've uncovered as well as the overall language implementation is dubious, but you need partial evaluation if not for performance - then for language interop.

If you are really serious about ML to the point of dedicating your life to it, then you should be able to do something like make a ML library from start to finish. That means being able to handle GPU and other kinds of hardware in an elegant fashion.

I want to trivialize this effort. I need to bring out the compositionality of ML to the point where I am doing is not a big deal. I can see here that big frameworks are not needed to do ML. I just need a single well written functional library.

And I must be realistic. Although I am aiming high, I can't realistically say that I am smarter than the guys at Deepmind. Or that I have bigger pockets or bigger computers than them.

All I can do is do the right thing, in order to make their efforts at 'saving' the world as expensive as possible. I must be humble and punish them for their hubris at going against the principles. And if I am to be humble, I should first look to myself.

If I am really serious, I should take all the advantages that I can get.

That means not abandoning my style. But it also means, not abandoning the ability to do top down reasoning in the quest for power.

10:35am. This life that I lead today is my punishment - I should have not gone to school 20 years ago and simply dove into programming. Had I a 14 year head start on programming, just where would I be now? All the hardship that I am going through today is because I lacked the strength of vision and the decisiveness back then. It might be luck of the draw, but rather than looking towards omnipotence, I should have looked at the earth.

10:40am. In this life, I am just revisiting the battles that I've already lost.

My vision and philosophy, and Spiral are gifts I wish I could send to my past self.

I regret him not having them.

But as a gift, v0.09 is not good enough. I need v0.2. And as a consolation, although I won't be able to send it to the past, I will be able to send it to the future.

I want people to benefit from v0.2 whether that be me or somebody following in my footsteps.

10:45am. Let me make the lightning strike twice. For the blessed future where I am not working on a language, but using it to make my own power instead.

Let me make that second great accomplishment in my life, and the third will take care of itself."