"1:20pm. Done with breakfast. Let me watch Re:Zero.
1:50pm. If the first ep is anything to go by, it will be just as good as the previous season. I'll definitely be watching it.

Now, focus me, focus.

2:30pm. Had to take a break, but I've been thinking about it during it.

Thoughts are still streaming through my mind, so I do not feel like talking about it just yet, but before I go to bed, let me write down what I am thinking currently.

Half an hour ago I said start, and if I am going to start, the place it makes sense to do so would be comments. That immediatelly raises trouble. Unlike in the previous parser where I filtered them out in the tokenizer, here I am keeping them around and that will cause issues for me if it turns out I have to add special cases to the language in order to deal with them.

So in order to simplify dealing with them...

```fs
    l : Tokenize.LineToken [] []
```

I think I will split this into regular tokens and comments. Comments have a property that they always come at the end so they will be easy to partition.

That will make them easy to associate to statements whenever I encounter them. I'll want that for variable tooltips.

I thought doing that only on the outer level, but smart dealing of the comments will make it possible to do that easily for the inner expressions as well.

2:40pm. Ok, that is about it for the comments. That is one thing.

The other part is - how exactly will I do the second part of the semantic highlighting pass?

2:50pm. ...For the second part I do not necessarily want a lot, but one thing I do want is to differentiate value and type scopes. I want the type variables to have distinct coloring.

Other thing I'd want is for..

```
inl x = From: 1 nearTo: 2
```

I'd want this to be recognized as a function call and highlighted specifically.

2:55pm. Now, comments are easy, but how exactly do I deal with this?

I know how to parse a token sequence into an AST, but how exactly do I parse this so the original sequence is enriched.

3:05pm. Also...

```
inl f (left: right:) = ...
```

If there is one thing I like, that is punning. I'd like the `left:` and `right:` here to be highlighted as variables.

Furthermore, I want the same thing in records.

```
inl f {a b c} = ...
```

I want `a`, `b`, `c` to be highlighted as variables here. And...

```
inl f {a=q b=w c=e} = ...
```

I want `a`, `b`, `c` to be highlighted as symbols here.

3:15pm. It is slowly coming to me. Let me take off to bed for a bit."