"2:40pm. Though I've been having breakfast, this whole time I've been thinking about join point specialization and renaming.
Conclussion:

I am going to get rid of hash consing as it serves little purpose. Instead for the second join point key I will just use the hash nodes. And during renaming, I will change variable and functions to tags. The way I will do that is if I have a tuple of say...

```
var 2 : i32, var 2 : i32, var 2 : i32
```

I will rename this to...

```
var 0 : i32, var_tag 0, var 1 : i32
```

I'll do a similar thing for functions. Now that I've figured out how to do compilation of mutually recursive function blocks everything is much easier for me. The reason for the thing above is to make it easier to use the unbuilt hashing and equality. The function one is especially important for this if I do not want exponential slowdowns.

2:45pm. Something great will come out of doing the language server.

Assuming I do not do specialization wholesale and do some boxing, I will be able to reuse the results of past compilation. I'll have the first part of the join point key be a weak reference through the weak conditional table.

2:50pm. This will give me incremental compilation ability even in the partial evaluator. That sort of thing is to die for. I'll even be able to extend this to the codegen.

v0.2 is going to come out as good as a single person can humanly make it.

Damn it, I meant to start work on the parser and I did, but now thoughts of specialization are taking over. Let me step away from the screen for a bit. I might as well do the chores here and take a nap after that.

I'll get it out of my system and into code eventually.

2:55pm. Actually since I will never be decompressing join point keys, I can get away with tagging everything. That would be the easiest thing to implement and would be the most efficient for hashing, equality and comparison."