"3:05pm. Done with chores. Let me go to bed again. I will Think.
5:05pm. I've thought. I definitely do not feel like programming now, but let me write down my thoughts.

First of all, incremental compilation of the partial evaluation pass is impossible. Prototypes couple everything together. It might be possible to eliminate some of that cost using complex analysis, but it would be to costly in terms of developer effort. And possible might not be worth it in the end. It is not even worth considering.

It would be possible to do incremental compilation without prototypes, but the language would be too poor to bother using in that case.

I keep thinking how to improve compilation speeds.

I had the idea of deferring compilation at join points - this would require a specialized runtime for Spiral, but this would only be a benefit on code that is never run, but is partially evaluated.

I do not think this situation would occur that often in practice.

The that leaves one last thing - forget partial evaluation and just do evaluation. Just interpret the AST like a dynamic language would. And leave enough information so that at join points that lead to Cuda or other such things you can just compile the method on the fly.

This would work well.

It would be the easiest way to speed up Spiral. It would give the fastest possible compilation. One issue is that macros would not work in that case. This method would require some integration with the target platform.

But once again, there are dynamic languages on the .NET platform, so this is quite realitic as a plan. I just haven't dove properly into it yet. In 2017 I did put a decent amount of effort into making use of .NET reflection for .NET types under the hood, but it turned out to be difficult to deal with so I scrapped that idea.

5:15pm. Ok, one interesting thing. I've completely forgotten that I've decided to associate nominal names with symbols...

Actually, no. Forget that. Since I've decided that peval...

Hmmm, actually let me go with the idea. It would simply destructuring of nominals.

...Ah wait. If the nominal symbol is the identity, then it really would be hard to shadow them.

Shadowing is allowed in Spiral, and I do not want that a nominal declared in some library somewhere ever conflicts with user code.

5:25pm. At any rate, not doing incremental partial evaluation will simplify memory management.

If I were doing it incrementally, how would I get rid of dead join points? I have no idea.

5:30pm. I'll have nominals ids have their own namespace. The same goes for prototypes ids.

The symbol dictionary is one consideration I need to think about.

I'll use an immutable map for it. Since the prepass will work on a block by block basis, I do not want to copy that dictionary. Though I could keep track of the additions and undo them, a map would be easier.

5:35pm. No forget deciding that now. I'll do that when the time comes.

Regardless of how thing turn out, in the long term the dual language strategy will be the dominant one. It sort of already is given the dominance of Java and C#, but their design is just so poor. Spiral is much better designed for that kind of thing.

5:40pm. What do I need to do? Bring my vision to fruition.

Today've decided on a few things like how to do specialization and deal with nominals.

I wanted to do some serious work on the parser, but the above thoughts hijacked my mind.

5:45pm. Tomorrow, I think I will leave my compilation concerns aside and try following the plan. Do the parser.

It won't be that hard. It is mostly done anyway in the prototype and I just have to adapt the code.

I have to believe in join points. They are the most realistic solution to all kinds of compilation issues that plague the current crop of languages. The version of Spiral I am doing is whole-program optimizing, but v0.3 will no doubt allow switching between various compilation modes using join points.

5:50pm. I have to believe that v0.2 will be enough to get me to that next level. That it will give me the power to make those excellent ensemble agents.

Alexis King argued against whole-program optimization, but Haskell is not going to do shit for me. Forget about language interop with it. In order to write the kind of code that I want, I need Spiral period.

Hell, I will decide it! I will state it here and now!

If it turns out that the programs grow so large that whole program optimization is killing productivity - I will implement a straightforward interpreter integrated with .NET for Spiral.

I just have to decide to that if it comes to that.

For now, forget it. I have everything else to deal with first.

5:55pm. I can sort of imagine it - I think the way agents will program will be different. They will have a far greater degree of integration with the language they are using. It won't be syntax based like us humans are doing. They'll probably work on ASTs directly represented in their mind. Coupled with their speeds that will lead to entirely different language design constraints.

That will be then and this is now.

For the kind of programs I want to use Spiral for - long running optimization processes, it will be very hard to beat what I am trying to do with Spiral v0.2. I do not have to worry about it. Spiral v0.09 was few secs for a few kLOC, and v0.2 will be at least 10 times that. Just swapping immutable maps for arrays, and adding the prepass would be worth that much on its own. Once I add the top level inlining, it will be even more.

Under 10s for 100k LOC will be an acceptable level of performance for anything I might do personally.

It might not be a level acceptable for huge corporate projects, but it will be for anything a single person can create. Honestly, if C++ can be viable in the real world, then so can Spiral.

I won't worry about it.

6:05pm. And now that I've decided this, that means I can get rid of the slow weak conditional table (in favor in vanilla dict) everywhere and not think about hash consing. Merely just memoizing hashes in join point keys will be enough.

Tomorrow, I will definitely focus on the parser. I've probed and tested my opponent enough. It is time that I switch to aggressively attacking the problems in front of me. I will deal with the parsing, prepass, partial evaluation and codegen passes. I will deal with editor support. Then I will have v0.2 completed.

During 2021 I will spend the time doing the finishing touches for editor support and package management, but otherwise I expect to be fully using the language.

In fact 2021 is too late for my taste.

Is there any reason why I should not already start using it next month? I think not.

I could do a large amount of work in 2.5 weeks assuming I can sit down and start writing.

6:10pm. Let me stop here for the day. That SAO thread made me want to check out the novels and the manga for GGO so I will get to that. I really liked GGO. I wanted to like regular SAO as well, but it got too boring and cringy for me past the first half of S1 so I dropped it."