"5:20pm. I think I'll just call it a day here. I could do the union parser, but I really should be fresh for the next part. Right now I am tempted to leave nominals and prototypes for later, but that is not the right mindset to go here.
Unlike in January, I should take the time and put in all the features that I had in mind. That is the way to go here.

5:25pm. Then I should work on editor support. Everything should be given my full attention and care.

Actually, let me do one last thing for the day.

```fs
    let pat_type =
        pipe2 body (opt (skip_op ":" >>. fun d -> root_type false (d.is_top_down = false) d.is_top_down d))
```

Note how I call `root_type` here. Where applicable, I should be using the same pattern.

```fs
        let case_typecase =
            let clauses d =
                let bar = bar (col d)
                (optional bar >>. sepBy1 (root_type true false false .>>. (skip_op "=>" >>. next)) bar) d
```

Ok, typecase is one place where things work well.

```fs
        let case_typecase =
            let clauses d =
                let bar = bar (col d)
                (optional bar >>. sepBy1 (root_type true false false .>>. (skip_op "=>" >>. next)) bar) d

            pipe2 (skip_keyword SpecTypecase >>. root_type false true false .>> skip_keyword SpecWith) clauses
                (fun a b -> RawTypecase(a,List.toArray b))
```

But I really should be checking whether the phase is top down.

```fs
        let case_typecase =
            let clauses d =
                let bar = bar (col d)
                (optional bar >>. sepBy1 (root_type true false false .>>. (skip_op "=>" >>. next)) bar) d

            range ((skip_keyword SpecTypecase >>. root_type false true false .>> skip_keyword SpecWith) .>>. clauses)
            >>= (fun (r, (a, b)) d ->
                if d.is_top_down then Error [r,TypecaseNotAllowed]
                else Ok(RawTypecase(a,List.toArray b))
                )
```

This is good. Let me clean up this a little.

5:40pm.

```fs
| "`" -> (((read_small_var |>> RawTVar) <|> (rounds (fun d -> root_type false (d.is_top_down = false) d.is_top_down d))) |>> RawType) d
```

This is good.

```fs
        pipe2 (tdop Int32.MinValue)
            (opt (indent (i-1) (<=) (skip_op ":") >>. indent i (<=) (fun d -> root_type false (d.is_top_down = false) d.is_top_down d)))
            (fun a -> function Some b -> RawAnnot(a,b) | _ -> a)
            d
```

And here in the operators. I had some trouble with parentheses with the snipped before this one. Clearly a sign that it is time to wrap things up for the day.

Now if only the router would stop breaking up, I'll be able to push this to Github."