"9:45am. Today it seems I stayed in bed longer than usual. I played Project Brutality 3.0 until late, so far I like it more than all the other Doom mods that I've tried. It is great.
Lesson number one from my life - channel your will into where it is possible. You benefit nothing from taking on boredom and humiliation in the long run. The story of the Inspired was inspiring to me enough to accept the curse of labor. And that is pretty much it.

I need to do this thing. I need to complete Spiral v0.2, and then get my hands on those neurochips. When I do, I'll try making the poker agent again. If I can go from 1k hands per second to 1M or more, then that will be a game changer either way. I'll implement both the game and the agent on that device.

GPUs are just excedingly poorly suited for this sort of thing. They were useful enough as a motivator, but just throughput will not cut it. I need latency as well.

One thing has been proven to me - I do not actually have any capability to anticipate how well arbitrary models and algorithms will perform in practice. But one thing I can anticipate is how my perception of them will change once they start to work. Ultimately, the perfomance regime you are operating becomes your reality. And I need to break the prison imposed on me by GPUs.

I do not really know once I have the agent whether I will use it to make real life money, but one thing I am sure is that I want to make it. I need to get good at this sort of thing.

I am just so tired. I have been tired from 2015 onwards, so me saying that does not mean much, but I haven't been able to get the creative release I so wanted in RL yet. If I could just tap into it, I get closer to the essence of what I want to do. If I can just hit this goal, I'll be able to put all my failures up to now behind me.

I want a real success in my life. Just making a language is not enough - I want to use it in a domain where no other languages can exceed it.

10:05am. Let me slack for a while and then I will start. Today I will begin work on integrating the block parser with the language server.

10:35am. Done slacking. Let me resume the run.

...Now, I said I would do this part relatively quickly so I can move on to practical things, but I think I will stick to this for a good long while. Even before typechecking and the prepass, I want to get all the editor related stuff out of the way. That will take me a while.

What I am saying is that I got the urge to do this. I'll go beyond just semantic tokenization and do all the prepass related highlighting done as well. At that point the editor support should be in a pretty good shape.

10:45am. It does not matter if this takes me an entire month. I should take this on and feel satisfied after it is done. Thankfully unlike the parser which was one huge intervowen thing that I've finally managed to simplify, this auxiliary stuff is a series of different things.

This part and the typechecking parts are the pieces that will trully be new to Spiral.

The prepass itself will only have compilation of mutually recursive functions added to it, and the partial evaluator will only need to tackle prototypes as its new responsibility. Everything else with regards to them is ground that I've covered and dealt with before. They should go smoothly. The codegen is barely worth mentioning - it is the most straightforward of all passes and nothing new needs to be done there. It had some challenges back in 2017 and 2018, but the way to do proper code generation from the partially evaled AST has been a solved problem for a while now.

10:50am. Though it will be time consuming, at least the editor support stuff will be relatively well isolated from the rest, and the rewards of my work will be immediatelly visible on the screen. Unlike the parser whose benefits I've only been imagining in my head for a while now, here I will get more immediate feedback.

And now that I am mentioning it, that was only the case thus far - I can in fact get feedback from the parser now.

11am. That should be my priority for today.

I need to switch gears. It is time to refresh my memory of the server. Then I have to start thinking of how to plug the parser in.

I won't stress myself to do a lot today. Just do a little. I won't force motivation or try to do a large quantity of work as if the fate of the world depends on it. The quantity will come when it does.

11:15am. If needed, I'll just meditate today of what I need to do. Since I am in a ranting mood, let me think until I feel like doing."